sat: build src/saturn at O3, split files where it over-inlines
